+++ middleware/idempotency/interfaces_test.go
@@ -1,8 +1,6 @@
 package idempotency
 
-import (
-	"testing"
-)
+import "testing"
 
 func TestPackageLocalStoreAliasesPreserveV3PortsIdentity(_ *testing.T) {
 	var store Store
@@ -20,3 +18,6 @@ func requirePortsStore(Store) {}
 func requirePortsReservationReleaser(ReservationReleaser) {}
 
 func requirePortsReleasableStore(ReleasableStore) {}
+
+var _ func(ReleasableStore, Options) (*Middleware, error) = NewWithStore
+var _ func(Options) (*Middleware, error) = New